Nightscout settings can now be saved

+++ Diabetes/APIDataAccess/Models/NotificationSetting/UpdateNotificationSettingModel.cs
@@ -18,7 +18,7 @@ namespace APIDataAccess.Models.NotificationSetting
         }
 
         public int Id { get; set; }
-        public int Threshold { get; set; }
+        public float Threshold { get; set; }
         public ThresholdType ThresholdType { get; set; }
         public NotificationType NotificationType { get; set; }
         public string Note { get; set; }
